Component: Improved support for non standard language codes
- store filename language in the database to avoid need to change this
  when aliases are changed
- apply aliases when creating new languages
- avoid double calculation of language code when adding new file
