+++ weblate/formats/base.py
@@ -352,32 +352,32 @@ class TranslationFormat:
         raise NotImplementedError()
 
     @classmethod
-    def get_language_code(cls, code: str, language_format: Optional[str] = None):
+    def get_language_code(cls, code: str, language_format: Optional[str] = None) -> str:
         """Do any possible formatting needed for language code."""
         if not language_format:
             language_format = cls.language_format
         return getattr(cls, f"get_language_{language_format}")(code)
 
     @staticmethod
-    def get_language_posix(code: str):
+    def get_language_posix(code: str) -> str:
         return code
 
     @staticmethod
-    def get_language_bcp(code: str):
+    def get_language_bcp(code: str) -> str:
         return code.replace("_", "-")
 
     @staticmethod
-    def get_language_posix_long(code: str):
+    def get_language_posix_long(code: str) -> str:
         if code in EXPAND_LANGS:
             return EXPAND_LANGS[code]
         return code
 
     @classmethod
-    def get_language_bcp_long(cls, code: str):
+    def get_language_bcp_long(cls, code: str) -> str:
         return cls.get_language_posix_long(code).replace("_", "-")
 
     @staticmethod
-    def get_language_android(code: str):
+    def get_language_android(code: str) -> str:
         # Android doesn't use Hans/Hant, but rather TW/CN variants
         if code == "zh_Hans":
             return "zh-rCN"
@@ -389,7 +389,7 @@ class TranslationFormat:
         return sanitized.replace("_", "-r")
 
     @classmethod
-    def get_language_java(cls, code: str):
+    def get_language_java(cls, code: str) -> str:
         # Java doesn't use Hans/Hant, but rather TW/CN variants
         if code == "zh_Hans":
             return "zh-CN"
@@ -402,14 +402,13 @@ class TranslationFormat:
         return cls.get_language_bcp(code)
 
     @classmethod
-    def get_language_filename(
-        cls, mask: str, code: str, language_format: Optional[str] = None
-    ):
-        """Return full filename of a language file.
-
-        Calculated forfor given path, filemask and language code.
+    def get_language_filename(cls, mask: str, code: str) -> str:
         """
-        return mask.replace("*", cls.get_language_code(code, language_format))
+        Returns full filename of a language file.
+
+        Calculated for given path, filemask and language code.
+        """
+        return mask.replace("*", code)
 
     @classmethod
     def add_language(cls, filename, language, base):
