+++ weblate/trans/tests/test_component.py
@@ -754,8 +754,9 @@ class ComponentValidationTest(RepoTestCase):
         self.assertEqual(component.get_lang_code("Solution/Project/Resources.resx"), "")
         # Language aliases
         self.assertEqual(
-            component.get_lang_code("Solution/Project/Resources.xx.resx"), "cs"
+            component.get_lang_code("Solution/Project/Resources.xx.resx"), "xx"
         )
+        self.assertEqual(project.get_language_alias("xx"), "cs")
         self.assertRaisesMessage(
             ValidationError,
             "The language code for "
